Borough-level American Community Survey estimates

A dataset containing U.S. Census Bureau American Community Survey (ACS) estimates and margins of error of selected demographic, social, and economic variables for all boroughs in New York City. Variables are five-year estimates from the 2013-2017 ACS.

borough_acs_data

Format

A tibble with 5 rows and 35 variables:

geoid

Census Bureau GEOID

pop_total_est

Total population estimate; B01001_001

med_age_est

Median age estimate; B01002_001

med_age_moe

Median age margin of error; B01002_001

med_hhinc_est

Median annual household income estimate (dollars); B19013_001

med_hhinc_moe

Median annual household income margin of error (dollars); B19013_001

pop_white_est, pop_white_pct_est

Non-hispanic white population estimate; B03002_003

pop_white_moe, pop_white_pct_moe

Non-hispanic white population margin of error; B03002_003

pop_black_est, pop_black_pct_est

Non-hispanic black population estimate; B03002_004

pop_black_moe, pop_black_pct_moe

Non-hispanic black population margin of error; B03002_004

pop_asian_est, pop_asian_pct_est

Non-hispanic asian population estimate; B03002_006

pop_asian_moe, pop_asian_pct_moe

Non-hispanic asian population margin of error; B03002_006

pop_hisp_est, pop_hisp_pct_est

Hispanic, any race population estimate; B03002_012

pop_hisp_moe, pop_hisp_pct_moe

Hispanic, any race population margin of error; B03002_012

pop_ba_above_est, pop_ba_above_pct_est

Population 25 years or older with at least a Bachelor's degree estimate; B15003_022, B15003_023, B15003_024, B15003_025

pop_ba_above_moe, pop_ba_above_pct_moe

Population 25 years or older with at least a Bachelor's degree margin of error; B15003_022, B15003_023, B15003_024, B15003_025

pop_educ_denom_est

Population 25 years or older estimate; Denominator used to calculate Bachelor's and above percentage; B15003_001

pop_educ_denom_moe

Population 25 years or older margin of error; Denominator used to calculate Bachelor's and above percentage; B15003_001

pop_inpov_est, pop_inpov_pct_est

Population with income below poverty line estimate; B17021_002

pop_inpov_moe, pop_inpov_pct_moe

Population with income below poverty line margin of error; B17021_002

pop_inpov_denom_est

Population for whom poverty status is determined estimate; Denominator used to calculate in poverty percentage; B17021_001

pop_inpov_denom_moe

Population for whom poverty status is determined margin of error; Denominator used to calculate in poverty percentage; B17021_001

Source

https://factfinder.census.gov/faces/nav/jsf/pages/index.xhtml